+++ ChangeLog.md
@@ -8,6 +8,8 @@
   Rogallo. ([#336](https://github.com/davep/rogallo/pull/336))
 - Added a `SourceSource` command to save the source of the current document.
   ([#338](https://github.com/davep/rogallo/pull/338))
+- Fixed `ToggleView` not allowing you to view the source of a Gopher map.
+  ([#340](https://github.com/davep/rogallo/pull/340))
 
 ## v1.8.0